新增, 修改, 删除  openhelper

+++ appBase/src/main/java/com/zhang/zhangandroid/persistence/sqlite/DatabaseHelper.java
@@ -35,7 +35,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         String CREATE_CONTACTS_TABLE = "CREATE TABLE " + TABLE_CONTACTS + "("
-                + KEY_ID + " INTEGER PRIMARY KEY," + KEY_NAME + " TEXT,"
+                + KEY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT ," + KEY_NAME + " TEXT,"
                 + KEY_PH_NO + " TEXT" + ")";
         db.execSQL(CREATE_CONTACTS_TABLE);
     }
@@ -51,7 +51,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     }
 
     /**
-     *增删改查操作
+     * 增删改查操作
      */
 
     // 增加新的联系人
@@ -62,29 +62,37 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         values.put(KEY_NAME, contact.getName());
         values.put(KEY_PH_NO, contact.getPhoneNumber());
 
-        // 插入行
-        db.insert(TABLE_CONTACTS, null, values);
-        db.close(); // 关闭数据库的连接
+        if (db.isOpen()) {
+            // 插入行
+            db.insert(TABLE_CONTACTS, null, values);
+            db.close(); // 关闭数据库的连接
+        }
+
     }
 
     // 获取联系人
     Contact getContact(int id) {
         SQLiteDatabase db = this.getReadableDatabase();
 
-        Cursor cursor = db.query(TABLE_CONTACTS, new String[] { KEY_ID,
-                KEY_NAME, KEY_PH_NO }, KEY_ID + "=?",
-                new String[] { String.valueOf(id) }, null, null, null, null);
-        if (cursor != null)
-            cursor.moveToFirst();
+        if (db.isOpen()) {
+            Cursor cursor = db.query(TABLE_CONTACTS, new String[]{KEY_ID,
+                            KEY_NAME, KEY_PH_NO}, KEY_ID + "=?",
+                    new String[]{String.valueOf(id)}, null, null, null, null);
+            if (cursor != null)
+                cursor.moveToFirst();
 
-        Contact contact = new Contact(Integer.parseInt(cursor.getString(0)),
-                cursor.getString(1), cursor.getString(2));
+            Contact contact = new Contact(Integer.parseInt(cursor.getString(0)),
+                    cursor.getString(1), cursor.getString(2));
 
-        cursor.close();
+            cursor.close();
+            db.close();
 
-        return contact;
+            return contact;
+        }
+        return null;
+
     }
-   
+
     // 获取所有联系人
     public List<Contact> getAllContacts() {
         List<Contact> contactList = new ArrayList<Contact>();
@@ -92,25 +100,27 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         String selectQuery = "SELECT  * FROM " + TABLE_CONTACTS;
 
         SQLiteDatabase db = this.getWritableDatabase();
-        Cursor cursor = db.rawQuery(selectQuery, null);
+        if (db.isOpen()) {
+            Cursor cursor = db.rawQuery(selectQuery, null);
 
-        if (cursor.moveToFirst()) {
-            do {
-                Contact contact = new Contact();
+            if (cursor.moveToFirst()) {
+                do {
+                    Contact contact = new Contact();
 
-                // 下面是获取字段的标准写法
-                contact.setID(Integer.parseInt(cursor.getString(cursor.getColumnIndex(KEY_ID))));
-                contact.setName(cursor.getString(cursor.getColumnIndex(KEY_NAME)));
-                contact.setPhoneNumber(cursor.getString(cursor.getColumnIndex(KEY_PH_NO)));
+                    // 下面是获取字段的标准写法
+                    contact.setID(Integer.parseInt(cursor.getString(cursor.getColumnIndex(KEY_ID))));
+                    contact.setName(cursor.getString(cursor.getColumnIndex(KEY_NAME)));
+                    contact.setPhoneNumber(cursor.getString(cursor.getColumnIndex(KEY_PH_NO)));
 
-                // 放入集合中
-                contactList.add(contact);
+                    // 放入集合中
+                    contactList.add(contact);
 
-            } while (cursor.moveToNext());
+                } while (cursor.moveToNext());
+            }
+
+            cursor.close();
+            db.close();
         }
-
-        cursor.close();
-        db.close();
 
         return contactList;
     }
@@ -119,24 +129,31 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public int updateContact(Contact contact) {
         SQLiteDatabase db = this.getWritableDatabase();
 
-        ContentValues values = new ContentValues();
-        values.put(KEY_NAME, contact.getName());
-        values.put(KEY_PH_NO, contact.getPhoneNumber());
+        if (db.isOpen()) {
+            ContentValues values = new ContentValues();
+            values.put(KEY_NAME, contact.getName());
+            values.put(KEY_PH_NO, contact.getPhoneNumber());
 
-        //更新行
-        int update = db.update(TABLE_CONTACTS, values, KEY_ID + " = ?",
-                new String[]{String.valueOf(contact.getID())});
+            //更新行
+            int update = db.update(TABLE_CONTACTS, values, KEY_ID + " = ?",
+                    new String[]{String.valueOf(contact.getID())});
 
+            db.close();
+            return update;
+        }
 
-        return update;
+        return -2;
     }
 
     // 删除单个联系人
     public void deleteContact(Contact contact) {
         SQLiteDatabase db = this.getWritableDatabase();
-        db.delete(TABLE_CONTACTS, KEY_ID + " = ?",
-                new String[] { String.valueOf(contact.getID()) });
-        db.close();
+
+        if (db.isOpen()) {
+            db.delete(TABLE_CONTACTS, KEY_ID + " = ?",
+                    new String[]{String.valueOf(contact.getID())});
+            db.close();
+        }
     }
 
 
@@ -144,13 +161,17 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public int getContactsCount() {
         String countQuery = "SELECT  * FROM " + TABLE_CONTACTS;
         SQLiteDatabase db = this.getReadableDatabase();
-        Cursor cursor = db.rawQuery(countQuery, null);
-        int count = cursor.getCount();
+        if (db.isOpen()) {
+            Cursor cursor = db.rawQuery(countQuery, null);
+            int count = cursor.getCount();
 
-        // 先取出数据后再关闭cursor
-        cursor.close();
+            // 先取出数据后再关闭cursor
+            cursor.close();
+            db.close();
+            return count;
+        }
 
-        return count;
+        return 0;
     }
